+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Changed
 
+- Session - Get peers get also users of subgroups
 - Deployment - Fix interacting with non-free resources
 - Group - Bulk delete removed not needed
 - Session - Deprecate groups attribute in session and cookie